moved openai and clip encoders
