+++ pyserini/encode/_openai.py
@@ -30,6 +30,7 @@ org_key = '' if os.getenv("OPENAI_ORG_KEY") is None else os.getenv("OPENAI_ORG_K
 client = openai.OpenAI(api_key=api_key, organization=org_key)
 OPENAI_API_RETRY_DELAY = 5
 
+
 def retry_with_delay(func, delay: int = OPENAI_API_RETRY_DELAY, max_retries: int = 10, errors: tuple = (openai.RateLimitError)):
     def wrapper(*args, **kwargs):
         num_retries = 0
@@ -45,7 +46,8 @@ def retry_with_delay(func, delay: int = OPENAI_API_RETRY_DELAY, max_retries: int
                 raise e
     return wrapper
 
-class OpenAIDocumentEncoder(DocumentEncoder):
+
+class OpenAiDocumentEncoder(DocumentEncoder):
     def __init__(self, model_name: str = 'text-embedding-ada-002', tokenizer_name: str = 'cl100k_base', **kwargs):
         self.model = model_name
         self.tokenizer = tiktoken.get_encoding(tokenizer_name)
@@ -62,15 +64,42 @@ class OpenAIDocumentEncoder(DocumentEncoder):
         inputs = [embedding[:max_length] for embedding in inputs]
         return self.get_embeddings(inputs)
     
-class OpenAIQueryEncoder(QueryEncoder):
-    def __init__(self, model_name: str = 'text-embedding-ada-002', tokenizer_name: str = 'cl100k_base', device = None):
-        self.model = model_name
-        self.tokenizer = tiktoken.get_encoding(tokenizer_name)
+# class OpenAIQueryEncoder(QueryEncoder):
+#     def __init__(self, model_name: str = 'text-embedding-ada-002', tokenizer_name: str = 'cl100k_base', device = None):
+#         self.model = model_name
+#         self.tokenizer = tiktoken.get_encoding(tokenizer_name)
+#
+#     @retry_with_delay
+#     def get_embedding(self, text: str):
+#         return np.array(client.embeddings.create(input=text, model=self.model)['data'][0]['embedding'])
+#
+#     def encode(self, text: str, max_length: int = 512, **kwargs):
+#         inputs = self.tokenizer.encode(text=text)[:max_length]
+#         return self.get_embedding(inputs)
+
+
+class OpenAiQueryEncoder(QueryEncoder):
+    def __init__(self, encoder_dir: str = None, encoded_query_dir: str = None,
+                 tokenizer_name: str = None, max_length: int = 512, **kwargs):
+        super().__init__(encoded_query_dir)
+        if encoder_dir:
+            api_key = '' if os.getenv("OPENAI_API_KEY") is None else os.getenv("OPENAI_API_KEY")
+            org_key = '' if os.getenv("OPENAI_ORG_KEY") is None else os.getenv("OPENAI_ORG_KEY")
+            self.client = openai.OpenAI(api_key=api_key, organization=org_key)
+            self.model = encoder_dir
+            self.tokenizer = tiktoken.get_encoding(tokenizer_name)
+            self.max_length = max_length
+            self.has_model = True
+        if (not self.has_model) and (not self.has_encoded_query):
+            raise Exception('Neither query encoder model nor encoded queries provided. Please provide at least one')
 
     @retry_with_delay
     def get_embedding(self, text: str):
-        return np.array(client.embeddings.create(input=text, model=self.model)['data'][0]['embedding'])
+        return np.array(self.client.embeddings.create(input=text, model=self.model)['data'][0]['embedding'])
 
-    def encode(self, text: str, max_length: int = 512, **kwargs):
-        inputs = self.tokenizer.encode(text=text)[:max_length]
-        return self.get_embedding(inputs)
+    def encode(self, query: str, **kwargs):
+        if self.has_model:
+            inputs = self.tokenizer.encode(text=query)[:self.max_length]
+            return self.get_embedding(inputs)
+        else:
+            return super().encode(query)
